// carebridge-admin-nextjs/src/types/models/user.ts
import { Authority } from "./authority";
import { BaseEntity } from "./base-entity";
import { Biodata } from "./biodata";
import { DeviceInfo } from "./device-info";


export interface Privilege {
    id: number;
    name: string;
}

export interface Role extends BaseEntity {
    id: number;
    name: string;
    privileges: Privilege[];
}

export interface User extends BaseEntity {
    id: number;
    name: string;
    email: string;
    // password: string;
    loginAttempt: number;
    isLocked: boolean;
    lastLogin: Date;
    // deviceInfo: DeviceInfo[];
    // authority: Authority[];
    roles: Role[];
    biodata: Biodata;
}